requirements.txt
streamlit
pandas
numpy
gspread
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
httplib2
plotly
streamlit-option-menu